+++ app/app/incidents/[id]/page.tsx
@@ -1,0 +1,3 @@
+import IncidentDetail from "../../../incident-detail";
+
+export default function IncidentDetailPage(){return <IncidentDetail/>;}