+++ app/app/incidents/page.tsx
@@ -1,0 +1,3 @@
+import IncidentList from "../../incident-list";
+
+export default function IncidentsPage(){return <IncidentList/>;}